// src/mailing/mailing.config.d.ts
export declare const defaultSender = "TAWOL Technologies Limited";
export declare const baseUrl = "http://46.101.26.184:3499/email";
export declare const MailingConfigs: {
    HEADER_KEY: string;
    HEADER_VALUE: string;
    PLAIN_EVENT_URL: string;
    HTML_EVENT_URL: string;
    PLAIN_ATTACHMENT_URL: string;
    HTML_ATTACHMENT_URL: string;
};
//# sourceMappingURL=mailing.config.d.ts.map